cast to int

+++ main.java
@@ -34,7 +34,7 @@ class Floyds{
                 if((pred[j][(int)from] == (int)to) && myCities.containsKey((int)j)) {
                     double signValue = Math.round(best[(int)from][j] - distance);
                     System.out.print(myCities.get((int)j) + " ");
-                    System.out.println(signValue);
+                    System.out.println((int)signValue);
                 }
             }
         }
